ArcFoundry/requirements.txt
numpy
onnx==1.14.0
onnxruntime
onnx-graphsurgeon
onnxsim
pyyaml
